requirements.txt
PyQt6
PyOpenGL
PyOpenGL-accelerate
numpy
Pillow
